worked on checkmark 2 of pr #424, some fix up of scattering3d test class
